requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
cryptography==36.0.1
idna==3.3
prettytable==3.0.0
pycparser==2.21
PyMySQL==1.0.2
python-dotenv==0.19.2
requests==2.27.1
six==1.16.0
urllib3==1.26.8
wcwidth==0.2.5
